Moved macros to hvp user space ink tap dance code
