+++ users/hvp/rules.mk
@@ -1,0 +1,2 @@
+SRC += hvp.c
+SRC += tap_dances.c
