+++ users/hvp/tap_dances.h
@@ -1,0 +1,10 @@
+#pragma once
+#include "hvp.h"
+
+// Tap Dance Declarations
+enum tapdance_id
+{
+  TD1 = 0,
+  TD2,
+  TD3
+};
